requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
atomicwrites==1.4.1
attrs==23.2.0
Babel==2.15.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
coreapi==2.3.3
coreschema==0.0.4
coverage==7.5.3
cryptography==42.0.7
cssselect2==0.7.0
Django==5.0.6
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==24.2
django-sequences==3.0
django-simple-history==3.6.0
djangorestframework==3.15.1
docopt==0.6.2
drf-writable-nested==0.7.0
drf-yasg==1.21.7
drf-yasg2==1.19.4
et-xmlfile==1.1.0
execnet==2.1.1
factory-boy==3.3.0
Faker==25.3.0
gunicorn==22.0.0
html5lib==1.1
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.4
lxml==5.2.2
MarkupSafe==2.1.5
num2words==0.5.13
numpy==1.26.4
numpy-financial==1.0.0
openpyxl==3.1.2
oscrypto==1.3.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pluggy==1.5.0
psycopg2-binary==2.9.9
py==1.11.0
pycparser==2.22
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
pyparsing==3.1.2
pypdf==4.2.0
pypng==0.20220715.0
pytest==8.2.1
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-forked==1.6.0
pytest-xdist==3.6.1
python-bidi==0.4.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.32.2
requests-mock==1.12.1
responses==0.25.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
svglib==1.5.1
tinycss2==1.3.0
tomli==2.0.1
typing_extensions==4.12.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.1
setuptools
webencodings==0.5.1
xhtml2pdf==0.2.15 
redis 
django-redis
